Exclude common asset file types.
We don't need to extract all files from the archive. As we have a limited amount of space (about 300MB) we should only extract things we need. We _could_ whitelist, but doing so is tricky as it's js, php, but also XML and .elint* files, and potentially others. So in the meantime I think it's easier to blacklist certain extenions.