Add support for running Psalm
Add support for Psalm + Psalm Plugin WordPress. This is a bit different to other "linters" as this is more of a typechecker, but I think it fits the general paradigm well, and in project that use Psalm, we can at least run it in those contexts. In the future we might want to expand this to run a default config for things like taint analysis.

The major difference with Psalm is we need to install Composer dependencies as those are part of the type flow, and needed for Psalm scanning. At the moment I just installed production dependencies, however I'm not sure on the best route here. It wouldn't be uncommon to include your tests and the like in the Psalm config, in which case you're likely to also need the composer dev deps. I think it would be ok to not run Psalm against tests, but we'd need to somehow exclude them from any Psalm configuration.